# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Ocorrências",
        "description": "API para gerenciamento de ocorrências, despachos e comandos. Suporta integrações externas e sistemas internos.",
        "contact": {
            "email": "suporte@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8089",
            "description": "Servidor de API"
        }
    ],
    "paths": {
...
        "/api/occurrences": {
            "get": {
                "tags": [
                    "Occurrences"
                ],
                "summary": "Listar ocorrências",
                "description": "Retorna uma lista paginada de ocorrências com opções de filtro por status e tipo",
                "operationId": "listOccurrences",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por status da ocorrência",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "pending"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrar por tipo de ocorrência",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "incendio_urbano"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Número de itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 50,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de ocorrências retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OccurrenceListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado - API Key inválida ou ausente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Muitas requisições - Rate limit excedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/occurrences/{id}": {
            "get": {
                "tags": [
                    "Occurrences"
                ],
                "summary": "Obter detalhes de uma ocorrência",
                "description": "Retorna os detalhes completos de uma ocorrência específica, incluindo seus despachos",
                "operationId": "getOccurrence",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID único da ocorrência (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da ocorrência retornados com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OccurrenceResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ocorrência não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado - API Key inválida ou ausente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/occurrences/{id}/start": {
            "post": {
                "tags": [
                    "Occurrences"
                ],
                "summary": "Iniciar processamento de uma ocorrência",
                "description": "Inicia o processamento assíncrono de uma ocorrência. Retorna um comando que pode ser consultado posteriormente para verificar o status",
                "operationId": "startOccurrence",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID único da ocorrência (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Comando de início aceito para processamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Command"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ocorrência não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Requisição duplicada - Idempotency Key já utilizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado - API Key inválida ou ausente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    },
                    {
                        "idempotencyKey": []
                    }
                ]
            }
        },
        "/api/occurrences/{id}/resolve": {
            "post": {
                "tags": [
                    "Occurrences"
                ],
                "summary": "Resolver uma ocorrência",
                "description": "Marca uma ocorrência como resolvida. Retorna um comando que pode ser consultado posteriormente para verificar o status",
                "operationId": "resolveOccurrence",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID único da ocorrência (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Comando de resolução aceito para processamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Command"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ocorrência não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado - API Key inválida ou ausente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/occurrences/types": {
            "get": {
                "tags": [
                    "Occurrences"
                ],
                "summary": "Listar tipos de ocorrência disponíveis",
                "description": "Retorna todos os tipos de ocorrência disponíveis no sistema",
                "operationId": "getOccurrenceTypes",
                "responses": {
                    "200": {
                        "description": "Lista de tipos de ocorrência retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OccurrenceTypesResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado - API Key inválida ou ausente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/occurrences/status": {
            "get": {
                "tags": [
                    "Occurrences"
                ],
                "summary": "Listar status de ocorrência disponíveis",
                "description": "Retorna todos os status de ocorrência disponíveis no sistema",
                "operationId": "getOccurrenceStatuses",
                "responses": {
                    "200": {
                        "description": "Lista de status de ocorrência retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OccurrenceStatusesResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado - API Key inválida ou ausente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Occurrence": {
                "title": "Ocorrência",
                "description": "Modelo de dados de uma ocorrência",
                "properties": {
                    "id": {
                        "description": "Classe para definir schemas e configurações globais do Swagger/OpenAPI\n\nEsta classe contém todas as definições reutilizáveis de schemas, tags,\nsecurity schemes e informações gerais da API.",
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "external_id": {
                        "type": "string",
                        "example": "EXT-12345"
                    },
                    "type_code": {
                        "type": "string",
                        "example": "incendio_urbano"
                    },
                    "type_name": {
                        "type": "string",
                        "example": "Incêndio Urbano"
                    },
                    "type_category": {
                        "type": "string",
                        "example": "emergency"
                    },
                    "status_code": {
                        "type": "string",
                        "example": "pending"
                    },
                    "status_name": {
                        "type": "string",
                        "example": "Pendente"
                    },
                    "description": {
                        "type": "string",
                        "example": "Incêndio em prédio residencial"
                    },
                    "reported_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-02-01T14:32:00-03:00"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-02-01T14:32:00-03:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-02-01T14:32:00-03:00"
                    },
                    "dispatches": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Dispatch"
                        }
                    }
                },
                "type": "object"
            },
...
            "Command": {
                "title": "Comando",
                "description": "Modelo de dados de um comando assíncrono",
                "properties": {
                    "command_id": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440002"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "processing",
                            "completed",
                            "failed"
                        ],
                        "example": "completed"
                    },
                    "result": {
                        "description": "Resultado do comando quando status é 'completed'",
                        "type": "object",
                        "nullable": true
                    },
                    "error": {
                        "description": "Mensagem de erro quando status é 'failed'",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Error": {
                "title": "Erro",
                "description": "Modelo de resposta de erro",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Validation failed"
                    },
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid"
                    },
                    "errors": {
                        "description": "Detalhes dos erros de validação",
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "OccurrenceListResponse": {
                "title": "Lista de Ocorrências",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Occurrence"
                        }
                    },
                    "meta": {
                        "description": "Metadados de paginação",
                        "type": "object"
                    },
                    "links": {
                        "description": "Links de paginação",
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "OccurrenceResponse": {
                "title": "Ocorrência Detalhada",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Occurrence"
                    }
                },
                "type": "object"
            },
            "OccurrenceTypesResponse": {
                "title": "Tipos de Ocorrência",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "code": {
                                    "type": "string",
                                    "example": "incendio_urbano"
                                },
                                "name": {
                                    "type": "string",
                                    "example": "Incêndio Urbano"
                                },
                                "category": {
                                    "type": "string",
                                    "example": "emergency"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "OccurrenceStatusesResponse": {
                "title": "Status de Ocorrência",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "code": {
                                    "type": "string",
                                    "example": "pending"
                                },
                                "name": {
                                    "type": "string",
                                    "example": "Pendente"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "description": "Chave de API para autenticacao.",
                "name": "X-API-Key",
                "in": "header"
            },
            "idempotencyKey": {
                "type": "apiKey",
                "description": "Chave de idempotencia para operacoes de escrita.",
                "name": "Idempotency-Key",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "apiKey": []
        }
    ]
}